% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{calculate_ypll}
\alias{calculate_ypll}
\title{Calculate YPLL}
\usage{
calculate_ypll(
  dt,
  byvar = NULL,
  age_adjusted_output = TRUE,
  year_rle = NA,
  export_data_by_simno = FALSE
)
}
\description{
Calculate YPLL
}
