% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covid19d_cty}
\alias{covid19d_cty}
\title{covid19d_cty (2020 county-level covid-19 deaths)}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 87864 rows and 13 columns.
}
\usage{
covid19d_cty
}
\description{
covid19d_cty (2020 county-level covid-19 deaths)
}
\keyword{datasets}
