% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_county_info}
\alias{get_county_info}
\title{Get the total number of counties in the US (excluding Puerto Rico)}
\usage{
get_county_info()
}
\description{
Get the total number of counties in the US (excluding Puerto Rico)
}
