% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{get_covid_death_cty}
\alias{get_covid_death_cty}
\title{Provisional county-level COVID-19 deaths by quarter in 2020}
\usage{
get_covid_death_cty()
}
\description{
Provisional county-level COVID-19 deaths by quarter in 2020
}
