% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{get_mort_nation_state}
\alias{get_mort_nation_state}
\title{Provisional national and state-level COVID-19 and total deaths in 2020 and 2021}
\usage{
get_mort_nation_state(
  select_sex = "All Sexes",
  select_year = 2020,
  verbose = TRUE
)
}
\description{
Provisional national and state-level COVID-19 and total deaths in 2020 and 2021
}
