% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{get_provisional_le}
\alias{get_provisional_le}
\title{Get provisional life expectancy estimates in 2020}
\usage{
get_provisional_le(url = "https://www.cdc.gov/nchs/data/vsrr/VSRR10-508.pdf")
}
\description{
Get provisional life expectancy estimates in 2020
}
