% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{get_uninsure_pop}
\alias{get_uninsure_pop}
\title{Get County-Level Percent Uninsured in 2019 (SAHIE estimates)}
\usage{
get_uninsure_pop()
}
\description{
Get County-Level Percent Uninsured in 2019 (SAHIE estimates)
}
