% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{impute_sample_agg}
\alias{impute_sample_agg}
\title{imputed data agg}
\format{
An object of class \code{list} of length 2.
}
\usage{
impute_sample_agg
}
\description{
imputed data agg
}
\keyword{datasets}
