% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{impute_sample_qr}
\alias{impute_sample_qr}
\title{imputed data qr}
\format{
An object of class \code{list} of length 2.
}
\usage{
impute_sample_qr
}
\description{
imputed data qr
}
\keyword{datasets}
