% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{le}
\alias{le}
\title{life expectancy}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 7 rows and 5 columns.
}
\usage{
le
}
\description{
life expectancy
}
\keyword{datasets}
