% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{le2020}
\alias{le2020}
\title{le2020}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 7 rows and 12 columns.
}
\usage{
le2020
}
\description{
le2020
}
\keyword{datasets}
