% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mort2020}
\alias{mort2020}
\title{mort2020}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 87864 rows and 12 columns.
}
\usage{
mort2020
}
\description{
mort2020
}
\keyword{datasets}
