% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_age_adjusted_ypll}
\alias{plot_age_adjusted_ypll}
\title{bar charts}
\usage{
plot_age_adjusted_ypll(
  dt,
  byvar,
  year_rle = NA,
  usa_ypll_ls = NULL,
  panel_letter = NULL,
  usa_text_angle = 90,
  axis.text.x.size = 8,
  axis.text.x.angle = 30
)
}
\description{
bar charts
}
