% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_prop_ypll_by_age}
\alias{plot_prop_ypll_by_age}
\title{Plot proportion of total YPLL attributable to each age group}
\usage{
plot_prop_ypll_by_age(
  dt,
  byvar,
  n_age_breaks = 7,
  year_rle = NA,
  usa_ypll_ls = NULL,
  panel_letter = NULL,
  axis.text.y.size = 8
)
}
\description{
Plot proportion of total YPLL attributable to each age group
}
