% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_ratio_ypll_to_covid}
\alias{plot_ratio_ypll_to_covid}
\title{Plot ratio}
\usage{
plot_ratio_ypll_to_covid(
  dt,
  byvar,
  year_rle = NA,
  usa_ypll_ls = NULL,
  panel_letter = NULL,
  axis.text.y.size = 8
)
}
\description{
Plot ratio
}
