% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{policy}
\alias{policy}
\title{policy}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 12568 rows and 14 columns.
}
\usage{
policy
}
\description{
policy
}
\keyword{datasets}
