% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{set_age_breaks}
\alias{set_age_breaks}
\title{Specify the age breaks for the population}
\usage{
set_age_breaks(
  age_cut = c(18, 30, 40, 50, 65, 75, 85),
  max_age = Inf,
  return_cut = TRUE
)
}
\arguments{
\item{age_cut}{A vector of the numeric lower bounds of all age breaks.
The default is \code{c(18, 30, 40, 50, 65, 75, 85)}}

\item{max_age}{The maximum age allowed (numeric). The default is infinite (\code{Inf}).}
}
\value{
The output of this function is a \code{data.table} with the lower and upper bounds of the age break,
and a column of string age breaks.
}
\description{
Specify the age breaks for the population
}
