% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{std_pop_wgt}
\alias{std_pop_wgt}
\title{std_pop_wgt}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 7 rows and 3 columns.
}
\usage{
std_pop_wgt
}
\description{
std_pop_wgt
}
\keyword{datasets}
