% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{summarize_vars}
\alias{summarize_vars}
\title{Summarize variables to get mean or the 95\% intervals}
\usage{
summarize_vars(x)
}
\arguments{
\item{x}{A vector of values}
}
\description{
Summarize variables to get mean or the 95\% intervals
}
