% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{dl_us_standard_population}
\alias{dl_us_standard_population}
\title{Get 2000 US standard population}
\usage{
dl_us_standard_population(
  url = "https://seer.cancer.gov/stdpopulations/stdpop.singleagesthru99.txt"
)
}
\arguments{
\item{url}{A string of the url linking to the data of the 2000 US standard population.}
}
\description{
Get 2000 US standard population
}
