% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mort2020}
\alias{mort2020}
\title{mort2020 (2020 national and state-level covid-19 and total deaths)}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 371 rows and 6 columns.
}
\usage{
mort2020
}
\description{
mort2020 (2020 national and state-level covid-19 and total deaths)
}
\keyword{datasets}
